Bug 1672520 - Part 0 - Move all different classes defined in nsHttpConnectionMgr into separate files. r=necko-reviewers,kershaw
Differential Revision: https://phabricator.services.mozilla.com/D94392
