+++ netwerk/protocol/http/moz.build
@@ -93,7 +93,10 @@ UNIFIED_SOURCES += [
     "ClassifierDummyChannelChild.cpp",
     "ClassifierDummyChannelParent.cpp",
     "ConnectionDiagnostics.cpp",
+    "ConnectionEntry.cpp",
+    "ConnectionHandle.cpp",
     "DelayHttpChannelQueue.cpp",
+    "HalfOpenSocket.cpp",
     "Http2Compression.cpp",
     "Http2Push.cpp",
     "Http2Session.cpp",
